+++ packages/blocks/src/_common/components/ai-item/index.ts
@@ -1,3 +1,2 @@
 export * from './ai-item-list.js';
-export * from './config.js';
 export * from './types.js';
