+++ packages/playground/apps/default/specs-examples/ai/ai.ts
@@ -4,14 +4,12 @@ import {
   AffineAIPanelWidget,
   EdgelessEditorBlockSpecs,
   PageEditorBlockSpecs,
-  toolbarDefaultConfig,
 } from '@blocksuite/blocks';
 import { AffineFormatBarWidget } from '@blocksuite/blocks';
-import { html, type TemplateResult } from 'lit';
 import { literal, unsafeStatic } from 'lit/static-html.js';
 
-import { FormatBarAIButton } from './format-bar-ai-button';
-import { setupAIPanel } from './panel/setup';
+import { setupFormatBarEntry } from './entry/format-bar/setup-format-bar';
+import { setupSpaceEntry } from './entry/space/setup-space';
 
 export function getAISpecs() {
   const pageModeSpecs = PageEditorBlockSpecs.map(spec => {
@@ -31,28 +29,11 @@ export function getAISpecs() {
           disposableGroup.add(
             slots.widgetConnected.on(view => {
               if (view.component instanceof AffineAIPanelWidget) {
-                setupAIPanel(view.component);
+                setupSpaceEntry(view.component);
               }
 
               if (view.component instanceof AffineFormatBarWidget) {
-                const formatBar = view.component;
-                toolbarDefaultConfig(formatBar);
-                formatBar.addRawConfigItems(
-                  [
-                    {
-                      type: 'custom' as const,
-                      render(
-                        formatBar: AffineFormatBarWidget
-                      ): TemplateResult | null {
-                        const askAIButton = new FormatBarAIButton();
-                        askAIButton.host = formatBar.host;
-                        return html`${askAIButton}`;
-                      },
-                    },
-                    { type: 'divider' },
-                  ],
-                  0
-                );
+                setupFormatBarEntry(view.component);
               }
             })
           );
