+++ packages/playground/apps/default/specs-examples/ai/entry/format-bar/format-bar-ai-button.ts
@@ -1,12 +1,13 @@
 import { type EditorHost, WithDisposable } from '@blocksuite/block-std';
 import {
-  AIItemGroups,
   AIStarIcon,
   createButtonPopper,
   isInsidePageEditor,
 } from '@blocksuite/blocks';
 import { css, html, LitElement } from 'lit';
 import { customElement, property, query } from 'lit/decorators.js';
+
+import { AIItemGroups } from './config';
 
 @customElement('format-bar-ai-button')
 export class FormatBarAIButton extends WithDisposable(LitElement) {
