--- packages/playground/apps/default/specs-examples/ai/panel/response/index.ts
@@ -1,6 +1,0 @@
-import { getInsertBelowHandler } from './insert-below';
-import type { AIPanelResponseGenerator } from './types';
-
-export const AIResponseGenerators: AIPanelResponseGenerator[] = [
-  getInsertBelowHandler,
-];
