--- packages/playground/apps/default/specs-examples/ai/panel/response/types.ts
@@ -1,6 +1,0 @@
-import type { EditorHost } from '@blocksuite/block-std';
-import type { AffineAIPanelWidget } from '@blocksuite/blocks';
-
-export type AIPanelResponseGenerator = (
-  host: EditorHost
-) => (panel: AffineAIPanelWidget) => Promise<void>;
